Location struct lives in its own file now Added update booleans to query if data is fresh.
