+++ sdkproject/Assets/Mapbox/Unity/Location/DeviceLocationProvider.cs
@@ -80,12 +80,16 @@ namespace Mapbox.Unity.Location
 
 			while (true)
 			{
+				_currentLocation.IsHeadingUpdated = false;
+				_currentLocation.IsLocationUpdated = false;
+
 				var timestamp = Input.compass.timestamp;
 				if (Input.compass.enabled && timestamp > _lastHeadingTimestamp)
 				{
 					var heading = Input.compass.trueHeading;
 					_currentLocation.Heading = heading;
 					_lastHeadingTimestamp = timestamp;
+					_currentLocation.IsHeadingUpdated = true;
 				}
 
 				var lastData = Input.location.lastData;
@@ -94,6 +98,7 @@ namespace Mapbox.Unity.Location
 				{
 					_currentLocation.LatitudeLongitude = new Vector2d(lastData.latitude, lastData.longitude);
 					_currentLocation.Accuracy = (int)lastData.horizontalAccuracy;
+					_currentLocation.IsLocationUpdated = true;
 					_currentLocation.Timestamp = timestamp;
 					_lastLocationTimestamp = timestamp;
 				}
