+++ sdkproject/Assets/Mapbox/Unity/Location/Location.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 7811c4b9fcda44ce39a09952d7e0af04
+timeCreated: 1509996012
+licenseType: Pro
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
