+++ sdkproject/Assets/Mapbox/Unity/Location/TransformLocationProvider.cs
@@ -40,6 +40,8 @@ namespace Mapbox.Unity.Location
 			_currentLocation.LatitudeLongitude = _targetTransform.GetGeoPosition(_map.CenterMercator, _map.WorldRelativeScale);
 			_currentLocation.Accuracy = _accuracy;
 			_currentLocation.Timestamp = DateTime.UtcNow.Subtract(new DateTime(1970, 1, 1)).TotalSeconds;
+			_currentLocation.IsLocationUpdated = true;
+			_currentLocation.IsHeadingUpdated = true;
 		}
 	}
 }
